+++ src/App/frontend/src/layout/FileUpload/Tag/EditWindowComponent.tsx
@@ -14,7 +14,7 @@ import { useLanguage } from 'src/features/language/useLanguage';
 import { AttachmentFileName } from 'src/layout/FileUpload/FileUploadTable/AttachmentFileName';
 import { FileTableButtons } from 'src/layout/FileUpload/FileUploadTable/FileTableButtons';
 import { useFileTableRow } from 'src/layout/FileUpload/FileUploadTable/FileTableRowContext';
-import classes from 'src/layout/FileUploadWithTag/EditWindowComponent.module.css';
+import classes from 'src/layout/FileUpload/Tag/EditWindowComponent.module.css';
 import { useIndexedId } from 'src/utils/layout/DataModelLocation';
 import { useItemWhenType } from 'src/utils/layout/useNodeItem';
 import { optionFilter } from 'src/utils/options';
@@ -40,7 +40,7 @@ export function EditWindowComponent({
   options,
   isFetching,
 }: EditWindowProps): React.JSX.Element {
-  const { textResourceBindings } = useItemWhenType(baseComponentId, 'FileUploadWithTag');
+  const { textResourceBindings } = useItemWhenType(baseComponentId, 'FileUpload');
   const { langAsString } = useLanguage();
   const { setEditIndex } = useFileTableRow();
   const uploadedAttachment = isAttachmentUploaded(attachment) ? attachment : undefined;