+++ src/App/frontend/src/layout/FileUpload/Tag/hasTag.ts
@@ -1,0 +1,5 @@
+import type { CompExternal } from 'src/layout/layout';
+
+export function fileUploadHasTag(component: CompExternal<'FileUpload'>) {
+  return Boolean(component.options || component.optionsId || component.source);
+}