+++ src/Designer/frontend/packages/ux-editor/src/testing/layoutMock.ts
@@ -94,8 +94,6 @@ export const layoutMock: IInternalLayout = {
     [component1IdMock]: component1Mock,
     [component2IdMock]: component2Mock,
     [component3IdMock]: component3Mock,
-    [componentMocks[ComponentType.FileUploadWithTag].id]:
-      componentMocks[ComponentType.FileUploadWithTag],
     ComponentWithOptionsMock: componentWithOptionsMock,
   },
   containers: {
@@ -125,11 +123,7 @@ export const layoutMock: IInternalLayout = {
   order: {
     [baseContainerIdMock]: [container1IdMock, container2IdMock, 'ComponentWithOptionsMock'],
     [container1IdMock]: [component1IdMock],
-    [container2IdMock]: [
-      component2IdMock,
-      component3IdMock,
-      componentMocks[ComponentType.FileUploadWithTag].id,
-    ],
+    [container2IdMock]: [component2IdMock, component3IdMock],
   },
   customRootProperties: customRootPropertiesMock,
   customDataProperties: customDataPropertiesMock,
@@ -147,11 +141,7 @@ export const layout1Mock: ExternalFormLayout = {
       {
         id: container2IdMock,
         type: ComponentType.RepeatingGroup,
-        children: [
-          component2IdMock,
-          component3IdMock,
-          componentMocks[ComponentType.FileUploadWithTag].id,
-        ],
+        children: [component2IdMock, component3IdMock],
         dataModelBindings: { group: { field: 'some-path', dataType: '' } },
       },
       {
@@ -164,7 +154,6 @@ export const layout1Mock: ExternalFormLayout = {
         type: component2TypeMock,
       },
       componentMocks[ComponentType.FileUpload],
-      componentMocks[ComponentType.FileUploadWithTag],
       {
         id: 'ComponentWithOptionsMock',
         type: ComponentType.Checkboxes,